ex43-1

+++ ex43/Program.cs
@@ -14,28 +14,25 @@ void PrintData(string prefix, string value)
     Console.WriteLine(prefix + value);
 }
 
-double CoordinatesX(int k1, int b1, int k2, int b2)
+double CoordinatesX(double k1, double b1, double k2, double b2)
 {
     double x = (b2 - b1) / (k1 - k2);
     return x;
 }
 
-double CoordinatesY(int k1, int b1, double x)
+double CoordinatesY(double k1, double b1, double x)
 {
     double y = k1 * x + b1;
     return y;
 }
 
 Console.WriteLine("Задайте коэффициенты уравнения (y=k1*x+b1) для прямой 1 ");
-int k1 = ReadData("k1=");
-int b1 = ReadData("b1=");
+double k1 = ReadData("k1=");
+double b1 = ReadData("b1=");
 Console.WriteLine("Задайте коэффициенты уравнения (y=k2*x+b2) для прямой 2 ");
-int k2 = ReadData("k2=");
-int b2 = ReadData("b2=");
+double k2 = ReadData("k2=");
+double b2 = ReadData("b2=");
 double coordX = CoordinatesX(k1, b1, k2, b2);
-//double coordX=(b2 - b1) / (k1 - k2);
 double coordY = CoordinatesY(k1, b1, coordX);
-//double coordY= k1 * coordX + b1;
-//Console.WriteLine(coordX.ToString());
 PrintData("Координта Х точки пересечения ", coordX.ToString());
 PrintData("Координта Y точки пересечения ", coordY.ToString());
